backend/requirements.txt

flask==2.2.3
flask-cors==3.0.10
openai==1.3.0
python-dotenv==1.0.0
gunicorn==20.1.0
